Split js code into files
