+++ index.html
@@ -31,7 +31,7 @@
           </div>
         </div>
         <span class="nav-menu-element" id="navInventions">Винаходи</span>
-        <span class="nav-menu-element">Премії</span>
+        <span class="nav-menu-element" id="navAwards">Премії</span>
       </div>
       <i class="fa-solid fa-magnifying-glass nav-search"></i>
     </nav>
@@ -150,7 +150,7 @@
         </div>
       </div>
       <div class="awards">
-        <h2>Премії</h2>
+        <h2 id="awardsTitle">Премії</h2>
         <div class="awards-data">
           <div class="awards-data-main">
             <p>
@@ -200,6 +200,7 @@
         </div>
       </div>
     </main>
-    <script src="src/js/script.js"></script>
+    <script type="module" src="src/js/functions.js"></script>
+    <script type="module" src="src/js/script.js"></script>
   </body>
 </html>
