+++ src/js/functions.js
@@ -1,0 +1,12 @@
+export const hideHint = () => {
+  hint.style.opacity = "0%";
+  setTimeout(() => {
+    hint.style.visibility = "hidden";
+  }, 300);
+};
+
+export const scrollToElem = (element, offset) => {
+  const yOffset = -offset;
+  const y = element.getBoundingClientRect().top + window.pageYOffset + yOffset;
+  window.scrollTo({ top: y, behavior: "smooth" });
+};
