+++ src/styles/style.css
@@ -74,6 +74,10 @@ nav {
   width: 100%;
   height: 100px;
   border-bottom: solid 1px rgba(183, 183, 183, 0.5294117647);
+  -webkit-user-select: none;
+     -moz-user-select: none;
+      -ms-user-select: none;
+          user-select: none;
 }
 nav .nav-logo {
   font-size: 36px;
@@ -151,9 +155,15 @@ nav .nav-menu-element {
 nav .nav-menu-element:hover {
   color: #2eb4ea;
 }
-nav .nav-search {
+nav .nav-menu-element:hover:hover {
+  color: #2eb4ea;
+}
+nav .nav-menu .nav-search {
   font-size: 35px;
   cursor: pointer;
+}
+nav .nav-menu .nav-search:hover {
+  color: #2eb4ea;
 }
 
 main {
@@ -314,6 +324,9 @@ main .awards-data-publication {
 main .awards-data-publication h4 {
   font-size: 32px;
   font-weight: 500;
+}
+main .awards-data-publication ul {
+  margin-top: 20px;
 }
 main .awards-data-publication ul li {
   font-size: 24px;
